3 apis fixed: savedpost, interaction, allposts

+++ backend/src/main/java/com/twd/SpringSecurityJWT/repository/OurUserRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface OurUserRepo extends JpaRepository<OurUsers, Integer> {
+public interface OurUserRepo extends JpaRepository<OurUsers, Long> {
       Optional<OurUsers> findByEmail(String email);
 
       Optional<OurUsers> findById(Long id);
